Improvements of Marbles WeatherPlugin: * BBC stations have a priority now. * Display 20 instead of 10 weather stations.
svn path=/branches/marble/marble-gsoc-2009/; revision=988286

+++ src/plugins/render/weather/WeatherItem.cpp
@@ -55,7 +55,7 @@ class WeatherItemPrivate {
     }
     
     WeatherData m_currentWeather;
-    quint8 m_priority;
+    int m_priority;
     QAction *m_action;
     WeatherItem *m_parent;
     QSize m_temperatureSize;
